Added three column section.
Added Icons and given credit. Ensured that they are resized to load fast. Ensured that the three column seciton is repsonsive. Need to gather feedback and whether to use either the three column section or two column section. Need to add Slick carousel at below and receive feedback on how the horizonatal lines should fit on the page. Need to receive feedback on the footer - should the connect section and foooter be merged on desktop?